Final_P/requirements.txt
# Python Requirements for SEC505 IDS System
# Install with: pip install -r requirements.txt

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# Database
PyMySQL==1.1.0
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Data Visualization
matplotlib==3.8.2

# Network Analysis
# Note: dumpcap and CICFlowMeter need to be installed separately

# Utilities
python-dotenv==1.0.0

# Optional but recommended for production
# Werkzeug==3.0.1  # For password hashing
# gunicorn==21.2.0  # Production web server
# redis==5.0.1  # For advanced queue management
